feat: refactor ViewContributionHistory
